+++ spec/selectors/modal_spec.rb
@@ -2,90 +2,55 @@
 
 describe "modal selector" do
   before do
-    visit "/modal.html"
+    visit "/dialog.html"
   end
 
-  it "selects a dialog by title" do
-    modal = find(:element, :div, text: "Dialog content")
-    expect(find(:modal, "Dialog title")).to eq modal
-  end
-
-  it "selects an alertdialog by title" do
-    modal = find(:element, :div, text: "Alertdialog content")
-    expect(find(:modal, "Alertdialog title")).to eq modal
-  end
-
-  it "matches a dialog" do
-    modal = find(:element, :div, text: "Dialog content")
-    expect(modal).to match_selector :modal
-  end
-
-  it "selects by aria-label" do
-    modal = find(:element, :div, text: "Aria label content")
-    expect(find(:modal, "aria-label")).to eq modal
-  end
-
-  it "selects by partial aria-label" do
-    modal = find(:element, :div, text: "Aria label content")
-    expect(find(:modal, "aria-lab")).to eq modal
-  end
-
-  it "selects by exact aria-label" do
-    modal = find(:element, :div, text: "Aria label content")
-    expect(find(:modal, "aria-label", exact: true)).to eq modal
-  end
-
-  it "does not find exact aria-label with the wrong text" do
-    expect do
-      find(:modal, "aria-labe", exact: true)
-    end.to raise_error Capybara::ElementNotFound
-  end
-
-  it "selects by multiple aria-labelledby" do
-    modal = find(:element, :div, text: "Split label content")
-    expect(find(:modal, "split label")).to eq modal
-  end
-
-  it "selects by partial aria-labelledby" do
-    modal = find(:element, :div, text: "Split label content")
-    expect(find(:modal, "split la")).to eq modal
-  end
-
-  it "selects by exact aria-labelledby" do
-    modal = find(:element, :div, text: "Split label content")
-    expect(find(:modal, "split label", exact: true)).to eq modal
-  end
-
-  it "does not find by exact aria-labelledby with the wrong text" do
-    expect do
-      find(:modal, "split la", exact: true)
-    end.to raise_error Capybara::ElementNotFound
-  end
-
-  it "does not select a modal missing a role" do
-    expect(page).to have_no_selector :modal, "Missing role"
-  end
-
-  it "does not select a modal missing aria-modal=true" do
+  it "does not find a non-native dialog without aria-modal" do
     expect(page).to have_no_selector :modal, "Missing aria modal"
   end
 
-  it "selects a native modal" do
-    click_on "Open native"
-    modal = find(:element, :dialog)
-    expect(find(:modal, "Native modal heading")).to eq modal
+  it "finds a non-native dialog with aria-modal" do
+    expect(page).to have_selector :modal, "Modal title"
   end
 
-  it "does not selects an unopened native modal" do
-    expect(page).to have_no_selector :modal, "Native modal content", visible: :all
+  it "finds a non-native alertdialog with aria-modal" do
+    expect(page).to have_selector :modal, "Alertdialog title"
   end
 
-  describe "within_modal" do
+  it "does not find a closed native dialog" do
+    expect(page).to have_no_selector :modal, "Native dialog heading"
+  end
+
+  it "does not find a open native dialog opened as a non-modal dialog" do
+    click_on "Open native dialog"
+    expect(page).to have_no_selector :modal, "Native dialog heading"
+  end
+
+  it "finds a open native dialog opened as a modal dialog" do
+    click_on "Open native modal"
+    expect(page).to have_selector :modal, "Native dialog heading"
+  end
+
+  it "does not find a closed native dialog with role alertdialog" do
+    expect(page).to have_no_selector :modal, "Native alertdialog heading"
+  end
+
+  it "does not finds an open native dialog with role alertdialog opened as a non-modal" do
+    click_on "Open native alertdialog as dialog"
+    expect(page).to have_no_selector :modal, "Native alertdialog heading"
+  end
+
+  it "finds an open native dialog with role alertdialog opened as a modal" do
+    click_on "Open native alertdialog as modal"
+    expect(page).to have_selector :modal, "Native alertdialog heading"
+  end
+
+  describe "within_dialog with modal true" do
     it "limits to within the modal" do
-      within_modal "Dialog title" do
+      within_modal "Modal title" do
         expect(page).to have_text <<~TEXT.strip, exact: true
-          Dialog title
-          Dialog content
+          Modal title
+          Modal content
         TEXT
       end
     end